Revert "Add controllers"
