+++ src/main/java/ru/practicum/shareit/ShareItApp.java
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class ShareItApp {
 
-    public static void main(String[] args) {
-        SpringApplication.run(ShareItApp.class, args);
-    }
+	public static void main(String[] args) {
+		SpringApplication.run(ShareItApp.class, args);
+	}
 
 }
