+++ src/main/java/ru/practicum/shareit/booking/BookingController.java
@@ -1,0 +1,12 @@
+package ru.practicum.shareit.booking;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * TODO Sprint add-bookings.
+ */
+@RestController
+@RequestMapping(path = "/bookings")
+public class BookingController {
+}
