--- src/main/java/ru/practicum/shareit/booking/dto/BookingMapper.java
@@ -1,12 +1,0 @@
-package ru.practicum.shareit.booking.dto;
-
-import ru.practicum.shareit.booking.model.Booking;
-
-public class BookingMapper {
-    private BookingMapper() {
-    }
-
-    public static BookingDto toBookingDto(Booking booking) {
-        return new BookingDto(booking.getStart(), booking.getEnd(), booking.getItem(), booking.getBooker(), booking.getStatus());
-    }
-}
