--- src/main/java/ru/practicum/shareit/booking/dto/BookingSaveDto.java
@@ -1,16 +1,0 @@
-package ru.practicum.shareit.booking.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class BookingSaveDto {
-    private long itemId;
-    private LocalDateTime start;
-    private LocalDateTime end;
-}
