--- src/main/java/ru/practicum/shareit/booking/model/BookingState.java
@@ -1,10 +1,0 @@
-package ru.practicum.shareit.booking.model;
-
-public enum BookingState {
-    ALL,
-    CURRENT,
-    PAST,
-    FUTURE,
-    WAITING,
-    REJECTED
-}
