--- src/main/java/ru/practicum/shareit/exception/booking/BookingAccessDeniedException.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.exception.booking;
-
-public class BookingAccessDeniedException extends RuntimeException {
-    public BookingAccessDeniedException(String message) {
-        super(message);
-    }
-}
