--- src/main/java/ru/practicum/shareit/exception/booking/BookingConflictException.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.exception.booking;
-
-public class BookingConflictException extends RuntimeException {
-    public BookingConflictException(String message) {
-        super(message);
-    }
-}
