--- src/main/java/ru/practicum/shareit/exception/booking/BookingNotFoundException.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.exception.booking;
-
-public class BookingNotFoundException extends RuntimeException {
-    public BookingNotFoundException(String message) {
-        super(message);
-    }
-}
