--- src/main/java/ru/practicum/shareit/exception/booking/InvalidBookingPeriodException.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.exception.booking;
-
-public class InvalidBookingPeriodException extends RuntimeException {
-    public InvalidBookingPeriodException(String message) {
-        super(message);
-    }
-}
