--- src/main/java/ru/practicum/shareit/exception/item/ItemAccessDeniedException.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.exception.item;
-
-public class ItemAccessDeniedException extends RuntimeException {
-    public ItemAccessDeniedException(String message) {
-        super(message);
-    }
-}
