--- src/main/java/ru/practicum/shareit/exception/item/ItemNotFoundException.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.exception.item;
-
-public class ItemNotFoundException extends RuntimeException {
-    public ItemNotFoundException(String message) {
-        super(message);
-    }
-}
