--- src/main/java/ru/practicum/shareit/exception/item/ItemUnavailableException.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.exception.item;
-
-public class ItemUnavailableException extends RuntimeException {
-    public ItemUnavailableException(String message) {
-        super(message);
-    }
-}
