--- src/main/java/ru/practicum/shareit/exception/model/ErrorDetails.java
@@ -1,13 +1,0 @@
-package ru.practicum.shareit.exception.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@AllArgsConstructor
-@Getter
-@Setter
-public class ErrorDetails {
-    private String message;
-    private String details;
-}
