--- src/main/java/ru/practicum/shareit/exception/user/EmailAlreadyExistException.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.exception.user;
-
-public class EmailAlreadyExistException extends RuntimeException {
-    public EmailAlreadyExistException(String message) {
-        super(message);
-    }
-}
