--- src/main/java/ru/practicum/shareit/exception/user/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.exception.user;
-
-public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(String message) {
-        super(message);
-    }
-}
