+++ src/main/java/ru/practicum/shareit/item/dto/ItemDto.java
@@ -1,20 +1,7 @@
 package ru.practicum.shareit.item.dto;
 
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-@AllArgsConstructor
+/**
+ * TODO Sprint add-controllers.
+ */
 public class ItemDto {
-    private long id;
-    @NotBlank
-    private String name;
-    @NotBlank
-    private String description;
-    @NotNull
-    private Boolean available;
 }
