--- src/main/java/ru/practicum/shareit/item/dto/ItemMapper.java
@@ -1,17 +1,0 @@
-package ru.practicum.shareit.item.dto;
-
-import ru.practicum.shareit.item.model.Item;
-
-public class ItemMapper {
-    private ItemMapper() {
-
-    }
-
-    public static Item dtoToItem(ItemDto itemDto) {
-        Item item = new Item();
-        item.setName(itemDto.getName());
-        item.setDescription(itemDto.getDescription());
-        item.setAvailable(itemDto.getAvailable());
-        return item;
-    }
-}
