--- src/main/java/ru/practicum/shareit/item/dto/ItemWithBookingDateDto.java
@@ -1,25 +1,0 @@
-package ru.practicum.shareit.item.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Setter
-@Getter
-@AllArgsConstructor
-public class ItemWithBookingDateDto {
-    private long id;
-    @NotBlank
-    private String name;
-    @NotBlank
-    private String description;
-    @NotNull
-    private Boolean available;
-
-    private LocalDateTime lastBooking;
-    private LocalDateTime nextBooking;
-}
