--- src/main/java/ru/practicum/shareit/item/dto/ItemWithCommentsDto.java
@@ -1,27 +1,0 @@
-package ru.practicum.shareit.item.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-import ru.practicum.shareit.item.model.Comment;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class ItemWithCommentsDto {
-    private long id;
-    @NotBlank
-    private String name;
-    @NotBlank
-    private String description;
-    @NotNull
-    private Boolean available;
-    private LocalDateTime lastBooking;
-    private LocalDateTime nextBooking;
-    private List<Comment> comments;
-}
