--- src/main/java/ru/practicum/shareit/item/dto/UpdateItemDto.java
@@ -1,10 +1,0 @@
-package ru.practicum.shareit.item.dto;
-
-import lombok.Data;
-
-@Data
-public class UpdateItemDto {
-    private String name;
-    private String description;
-    private Boolean available;
-}
