+++ src/main/java/ru/practicum/shareit/item/model/Item.java
@@ -1,29 +1,7 @@
 package ru.practicum.shareit.item.model;
 
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import lombok.Data;
-import ru.practicum.shareit.user.model.User;
-
-@Data
-@Entity
-@Table(name = "items")
+/**
+ * TODO Sprint add-controllers.
+ */
 public class Item {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    @NotBlank
-    private String name;
-
-    private String description;
-
-    @NotNull
-    private Boolean available;
-
-    @ManyToOne(fetch = FetchType.EAGER)
-    @JoinColumn(name = "owner", referencedColumnName = "id")
-    private User owner;
 }
