--- src/main/java/ru/practicum/shareit/item/repository/CommentRepository.java
@@ -1,12 +1,0 @@
-package ru.practicum.shareit.item.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import ru.practicum.shareit.item.model.Comment;
-
-import java.util.List;
-
-@Repository
-public interface CommentRepository extends JpaRepository<Comment, Long> {
-    List<Comment> findByItemId(long itemId);
-}
