+++ src/main/java/ru/practicum/shareit/request/ItemRequest.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.request.model;
+package ru.practicum.shareit.request;
 
 /**
  * TODO Sprint add-item-requests.
