+++ src/main/java/ru/practicum/shareit/request/ItemRequestController.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.request.controller;
+package ru.practicum.shareit.request;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
