+++ src/main/java/ru/practicum/shareit/user/User.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.user;
+
+/**
+ * TODO Sprint add-controllers.
+ */
+public class User {
+}
