--- src/main/java/ru/practicum/shareit/user/dto/UserMapper.java
@@ -1,25 +1,0 @@
-package ru.practicum.shareit.user.dto;
-
-import ru.practicum.shareit.user.model.User;
-
-public class UserMapper {
-    private UserMapper() {
-
-    }
-
-    public static User dtoToUser(UserDto userDto) {
-        User user = new User();
-        user.setId(userDto.getId());
-        user.setName(userDto.getName());
-        user.setEmail(userDto.getEmail());
-        return user;
-    }
-
-    public static UserDto userToDto(User user) {
-        UserDto userDto = new UserDto();
-        userDto.setId(user.getId());
-        userDto.setName(user.getName());
-        userDto.setEmail(user.getEmail());
-        return userDto;
-    }
-}
