--- src/main/java/ru/practicum/shareit/user/model/User.java
@@ -1,20 +1,0 @@
-package ru.practicum.shareit.user.model;
-
-import jakarta.persistence.*;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import lombok.Data;
-
-@Data
-@Entity
-@Table(name = "users")
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-    @NotBlank
-    private String name;
-    @Email
-    @NotBlank
-    private String email;
-}
