+++ src/main/resources/application.properties
@@ -1,6 +1,18 @@
-spring.application.name=shareit
+spring.jpa.hibernate.ddl-auto=none
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQL10Dialect
+spring.jpa.properties.hibernate.format_sql=true
+spring.sql.init.mode=always
 
-spring.datasource.url=jdbc:postgresql://localhost:5432/java-shareit
-spring.datasource.username=postgres
-spring.datasource.password=
-spring.datasource.driver-class-name=org.postgresql.Driver
+logging.level.org.springframework.orm.jpa=INFO
+logging.level.org.springframework.transaction=INFO
+logging.level.org.springframework.transaction.interceptor=TRACE
+logging.level.org.springframework.orm.jpa.JpaTransactionManager=DEBUG
+
+#---
+# TODO Append connection to DB
+#---
+spring.config.activate.on-profile=ci,test
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.url=jdbc:h2:mem:shareit
+spring.datasource.username=test
+spring.datasource.password=test
